- Enhanced user-specific auction views and bidding functionality. - Revised schema and data scripts for stable database interaction. - Implemented validation for auction start and end dates. - Added custom error handling for improved user experience. - Refactored Auction and Bid classes for optimized data handling. - Updated Auction ID type to Long across controllers and services, including corresponding test refactoring.

+++ src/main/java/com/example/auctionservicesaplication/repository/AuctionRepository.java
@@ -5,14 +5,13 @@ import com.example.auctionservicesaplication.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 // Repository for managing Auction entities in the database.
 
 // The JpaRepository interface from Spring Data provides basic CRUD operations (Create, Read, Update, Delete).
 @Repository
-public interface AuctionRepository extends JpaRepository<Auction, BigDecimal> {
+public interface AuctionRepository extends JpaRepository<Auction, Long> {
     // Method to find auctions by the seller (user).
     List<Auction> findBySeller(User seller);
 }
